Address review feedback on monitor health (PR #5507)
- Fix integer overflow: cast shun_recovery_time_sec to int64_t before
  multiplying by 1000 to prevent overflow for large values (>24 days)
- Add docstring note about production lag==-2 override path (issue #959)
  not covered by this simplified extraction
- Remove trailing "fix plan" comment from test

+++ lib/MonitorHealthDecision.cpp
@@ -42,7 +42,7 @@ bool can_unshun_server(
 
 	// Mirror MyHGC recovery: compute max_wait_sec with timeout cap
 	int max_wait_sec;
-	if (shun_recovery_time_sec * 1000 >= connect_timeout_max_ms) {
+	if ((int64_t)shun_recovery_time_sec * 1000 >= connect_timeout_max_ms) {
 		max_wait_sec = connect_timeout_max_ms / 1000 - 1;
 	} else {
 		max_wait_sec = shun_recovery_time_sec;
